Fix tutorial for newer versions of the Mapbox API
The Mapbox JS was calling the older version of the Mapbox web API (v3).
The older version doesn’t require an access token to use.
The newer Mapbox api (v4) changes so that these acess tokens are
required.

I found that newer Mapbox accounts created after a certain date aren’t
available over the older web API, so anyone following this tutorial
and creating a new Mapbox account would have seen an error.

It’s a shame that the tutorial has broken because of this change, if
this happens again, I could look to see if I can update the tutorial
to use an open tile server instead as an alternative to Mapbox.